feat: move neo4j files in their entity packages
