+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/caseEvent/CaseEventNode.kt
@@ -1,5 +1,6 @@
-package io.whozoss.agentos.persistence.neo4j
+package io.whozoss.agentos.caseEvent
 
+import io.whozoss.agentos.caseFlow.CaseNode
 import org.springframework.data.neo4j.core.schema.Id
 import org.springframework.data.neo4j.core.schema.Node
 import org.springframework.data.neo4j.core.schema.Relationship
