+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/namespace/NamespaceNodeNeo4jRepository.kt
@@ -1,4 +1,4 @@
-package io.whozoss.agentos.persistence.neo4j
+package io.whozoss.agentos.namespace
 
 import org.springframework.data.neo4j.repository.Neo4jRepository
 import org.springframework.data.neo4j.repository.query.Query
