+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/user/UserNode.kt
@@ -1,7 +1,6 @@
-package io.whozoss.agentos.persistence.neo4j
+package io.whozoss.agentos.user
 
 import io.whozoss.agentos.sdk.entity.EntityMetadata
-import io.whozoss.agentos.user.User
 import org.springframework.data.neo4j.core.schema.Id
 import org.springframework.data.neo4j.core.schema.Node
 import java.time.Instant
