+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/user/UserNodeNeo4jRepository.kt
@@ -1,4 +1,4 @@
-package io.whozoss.agentos.persistence.neo4j
+package io.whozoss.agentos.user
 
 import org.springframework.data.neo4j.repository.Neo4jRepository
 import org.springframework.data.neo4j.repository.query.Query
@@ -14,7 +14,7 @@ interface UserNodeNeo4jRepository : Neo4jRepository<UserNode, String> {
     /**
      * Find all non-removed users.
      * Used by [Neo4jUserRepository.findByParent] which receives the fixed
-     * parent key [io.whozoss.agentos.user.UserRepository.USER_PARENT_KEY].
+     * parent key [UserRepository.USER_PARENT_KEY].
      */
     @Query(
         $$"""MATCH (u:User)
